requirements.txt
requests~=2.32.2
BeautifulSoup4~=4.12.3
pandas~=2.2.2
customtkinter~=5.2.2
pillow~=10.3.0
psycopg2